// @trycourier/react-designer/src/types/elemental.types.ts
export type IActionButtonStyle = "button" | "secondary" | "tertiary" | "link";